requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
pypdf
streamlit
chromadb
python-dotenv
langchain-text-splitters
supabase
pandas
plotly
scikit-learn